Changes data type which is used in DNN to float. Change StepSize parameter to new LearningRate parameter whose type is float. Change tolerance to 1e-6 for unit tests of neural network since changing data type to float reduces the precision.

+++ src/main/java/edu/snu/reef/dolphin/neuralnet/NeuralNetworkTask.java
@@ -53,11 +53,11 @@ public final class NeuralNetworkTask implements Task {
     public void evaluate(final INDArray output, final int expected) {
       ++totalNum;
 
-      double maxValue = output.getDouble(0);
+      float maxValue = output.getFloat(0);
       int maxIndex = 0;
       for (int i = 1; i < output.length(); ++i) {
-        if (output.getDouble(i) > maxValue) {
-          maxValue = output.getDouble(i);
+        if (output.getFloat(i) > maxValue) {
+          maxValue = output.getFloat(i);
           maxIndex = i;
         }
       }
@@ -77,8 +77,8 @@ public final class NeuralNetworkTask implements Task {
     /**
      * @return the prediction accuracy of model.
      */
-    public double getStats() {
-      return correctNum / (double) totalNum;
+    public float getStats() {
+      return correctNum / (float) totalNum;
     }
 
     /**
